Add files via upload
The objective of this code is to calculate poverty in UK, Denmark and Spain using the EU SILC dataset. The poverty rates are calculated according to both the US definition (absolute) and EU definition (relative). The code also produces some CrossTables to decompose what poverty can mean in terms of ability to meet expenses or other such variables.