Add files via upload
The code looks at unemployment in France using the French Labour Force Survey data. Using the ILO and INSEE definitions, the code calculates the official employment and unemployment rates. The code then looks at the evolution of structure of working-age population that is employed, unemployed and inactive between different quarters using CrossTables. Finally, the code looks at differences in unemployment rate on the basis of gender, age and highest qualification.