Add files via upload
The code has 2 parts. One uses the data from the French primary schools survey, and analyses progress of pupils between 1st and 3rd standard. Then it produces transition matrices between tests 1 and 2 and tests 2 and 3. The code then compares children born in Jan to children born in Dec and the evolution of their progress in France, further looking at differences arising between children of executives and blue collar workers as well. The next part of the code analyses data from PISA for France, Norway and Vietnam. It looks at the number of missing values, runs the PISA function to determine the country-level average scores in math, reading and science. The code then looks at differences amongst girls and boys and whether these differences are similar between countries. The code produces very informative and clear graphs to compare girls and boys, across countries with confidence intervals to see if the differences are statistically significant.